Finally integrate the complete formulas and use them at their full potential in the LTR double and add always algorithm (both monty and non monty). Using complete formulas allows to perform the double and add always loop in constant time with regard to the input scalar length (MSB position). This is more robust against side channel attacks as a complementary countermeasure to blinding.
NOTE: as of now, complete formulas are the *default* ones used in libecc
as they are safer wrt side channels and do not have a major impact on performance
(around 5 to 10 percent performance impact).
Non complete formulas are kept in the code as a conditional compilation
case with COMPLETE=0 to force them.