Add the countermeasure of adding a small multiple of the curve order (described in http://eprint.iacr.org/2011/232.pdf) to avoid leaking the nonce size in signature schems (ECDSA and so on). This countermeasure is only useful when non-complete formulas are used and BLINDING is not used.
This closes issue #10.

NOTE: beware that this countermeasure is only related to timing side channels
linked to the nonce size. This is absolutely not efficient against more
elaborate side channels (SPA and so on using microarchitectural/cache attacks
or EM/power scope based attacks). BLINDING=1 *must* be used in critical
contexts where such attacks are considered.