Migrate the Double and Add Always algorithm to Itoh et al. masked variant that is more robust against DPA (more specifically ADPA, or Address-bit DPA). This variant of the algorithm is always used since it has almost no impact on performance.
See "A Practical Countermeasure against Address-Bit Differential Power Analysis"
by Itoh, Izu and Takenaka for more information.

All in all, libecc has now the following approaches:

- SPA is thwarted using Double and Add Always, plus complete formulas
to avoid leaking point at infinity (avoid exceptions). Constant time
operations are (tentatively) used to limit leakage of different operations,
even though this task is very complex to achieve (especially in pure C).

- DDPA (Data DPA) is thwarted using blinding of the point (projective
coordinates) and of the scalar (with adding a random multiple of the
curve order with maximum entropy). Because of its major impact on
performance, blinding must be specifically tuned on by the used using the
BLINDING=1 switch.

- ADPA (Address-bit DPA) is limited using Itoh et al. Double and Add Always
masked variant.

All these countermeaures must, of course, be validated on the specific target
with leakage assessment. Because of the very nature of C code and CPU
microarchitectural details, it is very complex without such a leakage
assessment (that again depends on the target) to be sure that SCA protection
is indeed efficient.