Major rework of the scalar multiplication layer, with now the addition of the Montgomery Ladder in addition to the Double-and-Add-Always algorithm. Both algorithms are protected with Itoh's Address-bit DPA countermeasure.
Because of its inherent robustness against horizontal attacks, Montgomery Ladder
is used by default. Double-and-Add-Always can still be chosen through user
options in the makefile:

$ ADALWAYS=1 make

NOTE: a big improvement against the previous version of the library is that
now the MSB countermeasure provided in https://eprint.iacr.org/2011/232.pdf
is always used for scalars that are less than the square order (which is usually
the case in our library). This means that now everything is "constant time" and
independent of the scalar size even with incomplete formulas (except the point
at infinity handling for these of course).