Fix some unitialized values usage spotted by valgrind (credits @povauboin as exhibited in PR #15: the PR only concerns a small part of ECDSA verify, and this commit should also fix signature, public key generation and other algorithms).
As a side note, this "issue" should not be a security problem per se since
such uninitialized state is actually caught and handled by the prj_pt layer
using our MAGIC initialization values.
But better safe than sorry: it is cleaner to properly initialize our values
in the calling functions!