Workaround for windows on file rename
As the documentation for os.rename ( https://docs.python.org/3/library/os.html#os.rename ) says, "On Windows, if dst exists a FileExistsError is always raised."
Renaming the "curves_list.h.tmp" to "curves_list.h" on the fly fails on windows.

This modification allows the script to work properly on windows.

tested on windows 10 with python 3.9.0 and on linux (WSL) with python 3.8.2