sig: add gen_priv_key() member to signature algs structures
Previously, libecc expected the initialization of private keys from
a buffer. Most of the ECC signature algorithms defined in
ISO14888-3:2016 have a private key which is a random value from ]0,q[
but EdDSA (Ed448 and Ed25519) have a different kind of private key
(random value passed through a hash function). This preliminary work
adds as a private key generation function for all existing algs in
the lib and also define a generic version (drawing random values in
]0,q[) that is used for existing signature algs.

Signed-off-by: Ryad Benadjila <ryad.benadjila@gmail.com>
Reviewed-by: Arnaud EBALARD <arnaud.ebalard@ssi.gouv.fr>