return/utils: make error handling macros return error instead of performing while(1) loop
Most of libecc functions are void, i.e. do not return a status value because error
condition are not expected to happen. Nonetheless, MUST_HAVE() macros are used
everywhere to verify things that can not happen do not indeed happen. For instance,
a pointer parameter passed to a given function is not NULL, or a parameter has been
initialized prior to be passed to another function, etc.

When condition of MUST_HAVE() macro is not verified (which is not supposed to
happen at all), it was decided to handle that unexpected behavior in production
code using an infinite loop (e.g. a while(1)). This pattern has never been
an issue in practice as far as we can tell and had the nice effect of making
the code more readable by avoiding systematic handling of return values by the
caller:

	fp t0, t1, t2, t3, t4, t5;

        ...
	MUST_HAVE(param1 != NULL, ret, err);
        ...

	fp_mul(&t0, &in1->X, &in2->X);
	fp_mul(&t1, &in1->Y, &in2->Y);
	fp_mul(&t2, &in1->Z, &in2->Z);
	fp_add(&t3, &in1->X, &in1->Y);
	fp_add(&t4, &in2->X, &in2->Y);

	fp_mul(&t3, &t3, &t4);
	fp_add(&t4, &t0, &t1);
	fp_sub(&t3, &t3, &t4);
	fp_add(&t4, &in1->X, &in1->Z);
	fp_add(&t5, &in2->X, &in2->Z);

        ...

instead of

	fp t0, t1, t2, t3, t4, t5;

        ...
	MUST_HAVE(param1 != NULL, ret, err);
        ...

        ret = fp_mul(&t0, &in1->X, &in2->X);
	if (ret) {
	        goto err;
        }
	ret = fp_mul(&t1, &in1->Y, &in2->Y);
	if (ret) {
	        goto err;
        }
	ret = fp_mul(&t2, &in1->Z, &in2->Z);
	if (ret) {
	        goto err;
        }
	ret = fp_add(&t3, &in1->X, &in1->Y);
	if (ret) {
	        goto err;
        }
	ret = fp_add(&t4, &in2->X, &in2->Y);
	if (ret) {
	        goto err;
        }

        ...

While spending some time passing the library through various code analysis
tools, we ended up concluding that tools are not comfortable with infinite
loops. We then decided to take another look at the situation and ended up
introducing the following evolution in the library:

 - most function in the library now return an error value (0 on success,
   -1 on error)
 - the macro EG() (Error Goto) is introduced to keep the code readable
   while handling error conditions of functions
 - we enforce a single return path in all functions using an error label
 - uninit of nn, fp, etc is handled in a systematic way
 - the library will be free from dummy infinite loops

Previous code example will now look like this:

	int ret;
	fp t0, t1, t2, t3;
	t0.magic = t1.magic = t2.magic = t3.magic = 0;

	...

	MUST_HAVE(param1 != NULL, ret, err);
	...

	ret = fp_mul(&t0, &in->X, &in->X);                 EG(ret, err);
	ret = fp_mul(&t1, &in->Y, &in->Y);                 EG(ret, err);
	ret = fp_mul(&t2, &in->Z, &in->Z);                 EG(ret, err);
	ret = fp_mul(&t3, &in->X, &in->Y);                 EG(ret, err);
	ret = fp_add(&t3, &t3, &t3);                       EG(ret, err);

	ret = fp_mul(&out->Z, &in->X, &in->Z);             EG(ret, err);
	ret = fp_add(&out->Z, &out->Z, &out->Z);           EG(ret, err);
	ret = fp_mul(&out->X, &in->crv->a_monty, &out->Z); EG(ret, err);
	ret = fp_mul(&out->Y, &in->crv->b3_monty, &t2);    EG(ret, err);
	ret = fp_add(&out->Y, &out->X, &out->Y);           EG(ret, err);

	...

err:
	fp_uninit(&t0);
	fp_uninit(&t1);
	fp_uninit(&t2);
	fp_uninit(&t3);

	return ret

Even if we do not like at all the idea of using macros including goto and
do not intend to generalize their use, we decided to follow that path to
keep the code auditable/readable by humans.

Obviously, those changes will break the API of most functions. Examples
provided in the src/examples/ folder of the project have been updated
to document required evolution of the code for libecc user.

This first commit introduces a long series of directly and indirectly
related patches, which are all tagged return/...

Signed-off-by: Arnaud EBALARD <arnaud.ebalard@ssi.gouv.fr>
Signed-off-by: Ryad Benadjila <ryadbenadjila@gmail.com>