perf: introduce SMALL_STACK compile time option for targets with limited RAM
When the lib was released a few years ago, one could run ECDSA w/ SHA256 on
secp256r1 on a Cortex M0 target with 8KB of RAM. Recent tests on the library
showed that this was no more the case. By identifying and fixing the most
stack-hungry functions, it now possible to run it again on such target.

Some security features implemented since initial release (Itoh and al.
countermeasure when using Double-and-Add always is disabled) and some
mechanisms (eddsa) are not available in that mode.

Note that this commit also improve general stack usage when the SMALL_STACK
option is not set.

Signed-off-by: Ryad Benadjila <ryadbenadjila@gmail.com>
Signed-off-by: Arnaud EBALARD <arnaud.ebalard@ssi.gouv.fr>