Add non canonical u coordinate and small order public keys rejection.
Non canonical u coordinate is not fully compliant with RFC7748, but since
we use isogenies for X25519/448 computation it is complex to guess the result
of x-only Montgomery Ladder without computing it (for X25519 it would be easy
to pre-compute this in a table as there are not much possible non canonical u,
but this is not possible for X448: we leave this for now and will eventually fix
this in a clean way).

Small order public keys rejection, although not stated in RFC7748, is a defense
in depth counter measure as there is absolutely no reason for a peer to receive
such a public key in a clean Diffie-Hellman context!