Add GOSTR34-11-94 hash function in the deprecated algorithms as an example.
Code has been focused on readability more than optimization: this is an almost verbatim
transcription of the standard/RFC functions. There is room for major optimizations
using mixed 32 bits precomputed SBOXes and 'PSI' function factorization. However, since
this code is only here for testing purposes we consider that this is sufficient for
our needs.

Please *DO NOT* use this deprecated and weak hash function in production code, this
algoritm has mainly been included for compatibility and tests purposes!