Add a primitive to compute possible y coordinates on Weierstrass curve from x coordinate (using Fp square root).
Add a function for ECDSA public key recovery from the signature and the digest,
as this can be sometines useful.
WARNING: as explained in the function comment, this function have some probability
of failing because of the reduction modulo q of the 'r' in the ECDSA signature.
This probability should be very low for curves with cofactor = 1 (except for
generating a specific signature with biased r of course). For curves with
cofactor != 1, this function will probably fail.

Add the ECDSA public recovery attempt and comparison in the self tests.